'You are technically correct, sir. The best kind of correct.'
Without the int(), a quorum value of 2.5 would result when Redis::DistLock is
given 3 servers; thusly, you would need _all_ servers available in order to acquire a lock.

Obviously this is not desirable as one of the properties of Redis::DistLock is to avoid
individual Redis node failures where possible.